+++ app/Http/Controllers/ProjectController.php
@@ -10,7 +10,6 @@ use CDash\Model\Project;
 use CDash\Model\Repository;
 use CDash\Model\UserProject;
 use Illuminate\Http\JsonResponse;
-use Illuminate\Http\Response;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Gate;
 
@@ -148,16 +147,6 @@ final class ProjectController extends AbstractProjectController
 
         $pageTimer->end($response);
         return response()->json(cast_data_for_JSON($response));
-    }
-
-    public function overview(): Response
-    {
-        return response()->angular_view('overview');
-    }
-
-    public function manageOverview(): Response
-    {
-        return response()->angular_view('manageOverview');
     }
 
     public function ajaxDailyUpdatesCurl(): void
